feat: OpenAIRepository 프로토콜 생성
- 구체 타입의 이름을 DefaultOpenAIRepository로 변경
