+++ TravelGenie/TravelGenie/Domain/RepositoryInterface/OpenAIRepository.swift
@@ -1,0 +1,14 @@
+//
+//  OpenAIRepository.swift
+//  TravelGenie
+//
+//  Created by summercat on 2023/08/29.
+//
+
+import OpenAISwift
+
+protocol OpenAIRepository {
+    func request(
+        chatMessages: [ChatMessage],
+        completion: @escaping (Result<OpenAI<MessageResult>, OpenAIError>) -> Void)
+}
